renamed httphdr package to httputil. implemented SplitHostPort.
