+++ store/store.go
@@ -22,7 +22,7 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/orkunkaraduman/oscdn/fsutil"
-	"github.com/orkunkaraduman/oscdn/httphdr"
+	"github.com/orkunkaraduman/oscdn/httputil"
 	"github.com/orkunkaraduman/oscdn/ioutil"
 	"github.com/orkunkaraduman/oscdn/namedlock"
 )
@@ -530,7 +530,7 @@ func (s *Store) startDownload(ctx context.Context, baseURL, keyURL *url.URL) (do
 		return nil, err
 	}
 
-	if expires := httphdr.Expires(resp.Header, now); !expires.IsZero() && !expires.After(data.Info.ExpiresAt) {
+	if expires := httputil.Expires(resp.Header, now); !expires.IsZero() && !expires.After(data.Info.ExpiresAt) {
 		data.Info.ExpiresAt = expires
 	}
